+++ libraries/ml/src/multi-column-dimensionality-reduction/reduce-dimensionality.ts
@@ -43,6 +43,8 @@ export async function multiColReduceDimensionality(table: DG.DataFrame, columns:
       'must have the same length');
   }
 
+  const tv = grok.shell.tableView(table.name) ?? grok.shell.addTableView(table);
+
   const doReduce = async () => {
     const pg = DG.TaskBarProgressIndicator.create(
       `Initializing ${uiOptions.scatterPlotName ?? 'dimensionality reduction'} ...`);
@@ -56,8 +58,7 @@ export async function multiColReduceDimensionality(table: DG.DataFrame, columns:
           embedXCol = table.columns.add(DG.Column.float(embedColsNames[0], table.rowCount));
           embedYCol = table.columns.add(DG.Column.float(embedColsNames[1], table.rowCount));
           if (plotEmbeddings && !scatterPlot) {
-            scatterPlot = grok.shell
-              .tableView(table.name)
+            scatterPlot = tv
               .scatterPlot({...scatterPlotProps, x: embedColsNames[0], y: embedColsNames[1],
                 title: uiOptions.scatterPlotName ?? 'Embedding space'});
           }
@@ -80,8 +81,7 @@ export async function multiColReduceDimensionality(table: DG.DataFrame, columns:
         table.columns.add(DG.Column.float(embedColsNames[1], table.rowCount));
         let resolveF: Function | null = null;
         if (plotEmbeddings) {
-          scatterPlot = grok.shell
-            .tableView(table.name)
+          scatterPlot = tv
             .scatterPlot({...scatterPlotProps, x: embedColsNames[0], y: embedColsNames[1],
               title: uiOptions.scatterPlotName ?? 'Embedding space'});
           ui.setUpdateIndicator(scatterPlot.root, true);