+++ packages/Charts/src/viewers/tree/tree-viewer.ts
@@ -28,9 +28,12 @@ export class TreeViewer extends EChartViewer {
   sizeAggrType: DG.AggregationType;
   symbolSizeRange: [number, number] = [5, 20];
   aggregations: string[] = Object.values(DG.AGG).filter((f) => f !== DG.AGG.KEY && f !== DG.AGG.PIVOT);
+  aggregationsStr: string[] = Object.values({...DG.STR_AGG, ...DG.STAT_COUNTS});
   colorColumnName: string;
   colorAggrType: DG.AggregationType;
   selectionColor = DG.Color.toRgb(DG.Color.selectedRows);
+  applySizeAggr: boolean = false;
+  applyColorAggr: boolean = false;
 
   constructor() {
     super();
@@ -133,11 +136,21 @@ export class TreeViewer extends EChartViewer {
         p?.name === 'sizeAggrType' || p?.name === 'colorColumnName' || p?.name === 'colorAggrType') {
       if (p?.name === 'hierarchyColumnNames')
         this.chart.clear();
+      if (p?.name === 'colorColumnName' || p?.name === 'colorAggrType')
+        this.applyColorAggr = this.shouldApplyAggregation(this.colorColumnName, this.colorAggrType);
+      if (p?.name === 'sizeColumnName' || p?.name === 'sizeAggrType')
+        this.applySizeAggr = this.shouldApplyAggregation(this.sizeColumnName, this.sizeAggrType);
       this.render();
     } else
       super.onPropertyChanged(p, render);
   }
 
+  shouldApplyAggregation(columnName: string, aggrType: string): boolean {
+    const isColumnNumerical = this.dataFrame.columns.byName(columnName).matches('numerical');
+    const isAggregationApplicable = this.aggregationsStr.includes(aggrType);
+    return isColumnNumerical || (!isColumnNumerical && isAggregationApplicable);
+  }
+  
   onTableAttached(propertyChanged?: boolean) {
     const categoricalColumns = [...this.dataFrame.columns.categorical].sort((col1, col2) =>
       col1.categories.length - col2.categories.length);
@@ -169,14 +182,13 @@ export class TreeViewer extends EChartViewer {
   getSeriesData() {
     const aggregations = [];
 
-    if (this.sizeColumnName) {
+    if (this.sizeColumnName && this.applySizeAggr)
       aggregations.push({ type: <DG.AggregationType> this.sizeAggrType,
         columnName: this.sizeColumnName, propertyName: 'size' });
-    }
-    if (this.colorColumnName) {
+
+    if (this.colorColumnName && this.applyColorAggr)
       aggregations.push({ type: <DG.AggregationType> this.colorAggrType,
         columnName: this.colorColumnName, propertyName: 'color' });
-    }
 
     return [TreeUtils.toTree(this.dataFrame, this.hierarchyColumnNames, this.filter, null, aggregations)];
   }
@@ -187,11 +199,11 @@ export class TreeViewer extends EChartViewer {
 
     this.option.series[0].data = this.getSeriesData();
 
-    this.option.series[0]['symbolSize'] = this.sizeColumnName ?
+    this.option.series[0]['symbolSize'] = this.sizeColumnName && this.applySizeAggr ?
       (value: number, params: {[key: string]: any}) => utils.data.mapToRange(
         params.data.size, this.option.series[0].data[0]['size-meta']['min'],
         this.option.series[0].data[0]['size-meta']['max'], ...this.symbolSizeRange) : this.symbolSize;
-    if (this.colorColumnName)
+    if (this.colorColumnName && this.applyColorAggr)
       this.colorCodeTree(this.option.series[0].data[0]);
 
     this.chart.setOption(this.option);